event service impl

+++ ProEvents/Back/src/ProEvents.Domain/Entities/Entity.cs
@@ -2,6 +2,6 @@ namespace ProEvents.Domain.Entities
 {
     public class Entity
     {
-        public long Id { get; private set; }
+        public int Id { get; private set; }
     }
 }
